Add bot with Teo
