+++ discord.go
@@ -1,0 +1,52 @@
+package main
+
+import (
+	"fmt"
+	"os"
+	"os/signal"
+	"syscall"
+
+	"github.com/bwmarrin/discordgo"
+)
+
+var (
+	token = "OTgzNDk5MDMwNTIyOTAwNTIw.GNIN4u.uBh-QOjfZZASbo-pMh2ZayslszWJcVi6RSHZrM"
+)
+
+func ConnectToDiscord() {
+
+	dg, err := discordgo.New("Bot " + token)
+	if err != nil {
+		panic(err)
+	}
+
+	dg.AddHandler(messageCreate)
+
+	dg.Identify.Intents = discordgo.IntentsGuildMessages
+
+	err = dg.Open()
+	if err != nil {
+		panic(err)
+	}
+
+	fmt.Println("Bot is now running")
+	sc := make(chan os.Signal, 1)
+	signal.Notify(sc, syscall.SIGINT, syscall.SIGTERM, os.Interrupt) //, os.Kill)
+	<-sc
+
+	dg.Close()
+}
+
+func messageCreate(s *discordgo.Session, m *discordgo.MessageCreate) {
+	// ignore messages from bot himself
+	if m.Author.ID == s.State.User.ID {
+		return
+	}
+
+	// !Hello - bot will always accept challenges
+	if m.Content == "!Hello" {
+		s.ChannelMessageSend(m.ChannelID, "Hello World!")
+		return
+	}
+
+}
