+++ go.mod
@@ -1,0 +1,5 @@
+module los-concurrentes
+
+go 1.18
+
+require github.com/bwmarrin/discordgo v0.25.0 // indirect
